muokkausta

+++ Lisuke.php
@@ -3,6 +3,7 @@ require_once "tietokanta.php";
 
 $yhteys = getTietokanta();
 $id = $_GET["id"];
+
 $lisuke = NULL;
 if ($id != NULL)
 {
@@ -11,6 +12,7 @@ $kysely->execute();
 
 $lisuke = $kysely->fetch();
 }
+
 ?>
 <html>
     <head>
@@ -21,7 +23,7 @@ $lisuke = $kysely->fetch();
             <form action="Lisukelista.php" method="post">
             <input type="hidden" name="id" value="<?php echo $lisuke['lisukeid'] ?>" />
             Nimi <input type="text" name="nimi" value="<?php echo $lisuke['nimi'] ?>" />
-            Hinta<input type="integer" name="hinta" value="<?php echo $lisuke['hinta'] ?>" />
+            Hinta<input type="double" name="hinta" value="<?php echo $lisuke['hinta'] ?>" />
             <input type="submit" value="Lisää" />
          </form>
     </body>
